this_pointer.cpp
This program demonstrates the use of the this pointer in C++.

Key Concepts:

Every non-static member function has a hidden pointer called this which points to the object that invoked it.

Resolving ambiguity: When parameter names match data members (this->name = name).

Returning current object: Using return *this; allows method chaining.

this pointer is not available in static member functions.

Examples Covered:

Assigning values with this->.

Displaying object data.

Method chaining with return *this.